refactor: 프로젝트를 계층형 디렉터리 구조로 변환
- 코드의 역할에 따라 디렉터리를 분리하여 재구성하여 기능, 역할에 따른
  모듈화를 추구하고 프로젝트 구조에 대한 이해를 도움
