+++ api/src/main/java/burgermap/dto/member/MemberJoinDto.java
@@ -1,4 +1,4 @@
-package burgermap.members.dto;
+package burgermap.dto.member;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
